v0.8
This is to serve as initial commit of both client and server for my bachelor thesis with makeshift protoinstructions, makeshift categorisation of code and MOST IMPORTANTLY makeshift functionality that is YET TO BE PROVEN to work.
As far as concept goes, this is it and this is done in best faith.
Additional effort is to be made in far more rigorous instructions, far more rigorous code organisation (self-documenting and higher quality) and functionality is to be only adjusted very minutely unless something went wrong.

A side note:
"DBL" is yet to fire off, but this is believed to be result of "PIDW" overtax with too many blinks, not the the inability to detect in a calm "PIDW" into which "CACHE" is sectioned up.